Remove MatchWebFonts extension from mathjax config
MatchWebFonts is designed for pages that have mathematics within text
that is displayed using web fonts.  It can be configured to check for
web font arrivals so font size can be determined.
(http://docs.mathjax.org/en/latest/options/other/MatchWebFonts.html)

The way we use mathjax is we give it an equation to convert, together
with the font size.  We don't have a page, so this extension isn't
necessary.

The reason why we should remove it from the mathjax config is it appears
to have a memory leak the way we're using it.  Specifically, it's these
lines:

```
AJAX.timer.start(AJAX,["checkFonts",this,state.jax[this.id]],
                 CONFIG.fontCheckDelay,CONFIG.fontCheckTimeout);
```

https://github.com/mathjax/MathJax/blob/419b0a6eee7eefc0f85e47f7d4f8227ec28b8e57/unpacked/extensions/MatchWebFonts.js#L131-L132

The memory usage goes up with each math element we convert.