Fix mathjax format MathML / inline-TeX detection
Mathify looks for all `<math>` elements and elements with the
`data-math` attribute for conversion using mathjax.  All the `<math>`
elements should use "MathML" and all the elements with the `data-math`
attribute should use "inline-TeX" as the format when using the mathjax
API.  The old logic decides it's "inline-TeX" only if the math starts
with a backslash.  It works for math like `\text{W}" but not for
"E = mc^{2}".

So instead of checking if the math starts with a backslash, I changed it
to check if the math starts with a `^<([^:]+:)?math`, and if it does,
it's "MathML" and if it doesn't, it's "inline-TeX".  This works for
hsphysics (col10002 on content04).